Unit Changes

+++ DAL/Models/Stock.cs
@@ -5,9 +5,9 @@
         public int StockId { get; set; }        // Primary Key
 
         public int ProductId { get; set; }      // Foreign Key to Product
-        public int Quantity { get; set; }       // Current stock count
+        public decimal Quantity { get; set; }       // Current stock count
 
-        public int MinimumQuantity { get; set; }  // For low stock alert
+        public decimal MinimumQuantity { get; set; }  // For low stock alert
 
         // Navigation (optional, safe)
         public Product Product { get; set; }
