+++ DAL/Repository/Interface/IStockRepository.cs
@@ -7,15 +7,15 @@ namespace DAL.Repository.Interfaces
     {
         Stock GetById(int id);
         IEnumerable<Stock> GetAll();
-        IEnumerable<Stock> GetLowStock(int limit);
+        IEnumerable<Stock> GetLowStock(decimal limit);
 
         void Add(Stock stock);
         void Update(Stock stock);
         void Delete(int id);
 
         // Stock Management
-        void IncreaseStock(int productId, int qty);
-        void DecreaseStock(int productId, int qty);
+        void IncreaseStock(int productId, decimal qty);
+        void DecreaseStock(int productId, decimal qty);
         Stock GetByProductId(int productId);
         IDbContextTransaction BeginTransaction();
